add simple argument parsing
This commit adds argparsing and looks for a file in the first position. It
also adds updates to string formatting and prints all the words by
default (instead of the previous 10 word default).

One issue with `format()` is that it doesn't seem to take double-width
characters into account. This might occasionally cause some issues with
formatting.